pyproject.toml
[project]
name = "5-queens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "python-constraint>=1.4.0",
    "simpleai>=0.8.3",
]
